# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A REST API for swim training plan management with RAG capabilities",
        "title": "Swim Gen API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload and store a new user created swim training plan in the RAG system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation"
                ],
                "summary": "Donate a new training plan",
                "parameters": [
                    {
                        "description": "Training plan data",
                        "name": "plan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DonatePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan added successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/export-pdf": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate and download a PDF version of a training plan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training Plans"
                ],
                "summary": "Export training plan to PDF",
                "parameters": [
                    {
                        "description": "Training plan data to export",
                        "name": "plan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PlanToPDFRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF export response with URI",
                        "schema": {
                            "$ref": "#/definitions/models.PlanToPDFResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/generate-prompt": {
            "post": {
                "description": "Generate a prompt for the LLM based on the provided language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training Plans"
                ],
                "summary": "Generate a prompt for the LLM",
                "parameters": [
                    {
                        "description": "Request to generate a prompt",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GeneratePromptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated prompt response",
                        "schema": {
                            "$ref": "#/definitions/models.GeneratedPromptResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the health status of the API including database and vector store connectivity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Comprehensive health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HealthStatus"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/models.HealthStatus"
                        }
                    }
                }
            }
        },
        "/health-basic": {
            "get": {
                "description": "Returns a simple OK response for basic health monitoring",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Basic health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/query": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Query the RAG system for relevant training plans based on input",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training Plans"
                ],
                "summary": "Query training plans",
                "parameters": [
                    {
                        "description": "Query parameters",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query results",
                        "schema": {
                            "$ref": "#/definitions/models.RAGResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/scrape": {
            "get": {
                "description": "Scrape and import training plans from external websites",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Web Scraping"
                ],
                "summary": "Scrape training plans from web",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL to scrape training plans from",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scraping completed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request - Missing url parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/upsert-plan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing training plan if it belongs to the user, or insert a new one",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training Plans"
                ],
                "summary": "Update or insert a training plan into a user's history",
                "parameters": [
                    {
                        "description": "Request to upsert a training plan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpsertPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan ID of the upserted training plan",
                        "schema": {
                            "$ref": "#/definitions/models.UpsertPlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.DonatePlanRequest": {
            "description": "Request payload for donating a swim training plan to the system",
            "type": "object",
            "required": [
                "table"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "A comprehensive training plan for improving freestyle technique"
                },
                "language": {
                    "description": "Language specifies the language of the training plan",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Language"
                        }
                    ],
                    "example": "en"
                },
                "table": {
                    "description": "A structured training plan table containing exercise rows",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Row"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Advanced Freestyle Training"
                }
            }
        },
        "models.GeneratePromptRequest": {
            "description": "Request payload for generating a prompt for swim training plan creation",
            "type": "object",
            "required": [
                "language"
            ],
            "properties": {
                "language": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Language"
                        }
                    ],
                    "example": "en"
                }
            }
        },
        "models.GeneratedPromptResponse": {
            "description": "Response containing the generated prompt for swim training plan creation",
            "type": "object",
            "properties": {
                "prompt": {
                    "type": "string",
                    "example": "Generate a swim training plan for improving freestyle technique"
                }
            }
        },
        "models.HealthStatus": {
            "description": "Health status of the service and its components",
            "type": "object",
            "properties": {
                "components": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "schema_version": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.Language": {
            "type": "string",
            "enum": [
                "en",
                "de"
            ],
            "x-enum-varnames": [
                "LanguageEN",
                "LanguageDE"
            ]
        },
        "models.PlanToPDFRequest": {
            "description": "Request payload for exporting a training plan to PDF format",
            "type": "object",
            "required": [
                "description",
                "table",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "A comprehensive training plan for improving freestyle technique"
                },
                "horizontal": {
                    "description": "Horizontal indicates if the PDF should be in landscape orientation",
                    "type": "boolean",
                    "example": false
                },
                "language": {
                    "description": "Language specifies the language for the PDF content",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Language"
                        }
                    ],
                    "example": "en"
                },
                "large_font": {
                    "description": "LargeFont indicates if the PDF should use a larger font size",
                    "type": "boolean",
                    "example": true
                },
                "plan_id": {
                    "description": "PlanID identifies the training plan to be exported",
                    "type": "string",
                    "example": "plan_123"
                },
                "table": {
                    "description": "A structured training plan table containing exercise rows",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Row"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Advanced Freestyle Training"
                }
            }
        },
        "models.PlanToPDFResponse": {
            "description": "Response containing the URI to the generated PDF file",
            "type": "object",
            "properties": {
                "uri": {
                    "type": "string",
                    "example": "https://storage.googleapis.com/bucket/plans/plan_123.pdf"
                }
            }
        },
        "models.QueryRequest": {
            "description": "Request payload for querying swim training plans from the RAG system",
            "type": "object",
            "required": [
                "content",
                "method"
            ],
            "properties": {
                "content": {
                    "description": "Content describes what kind of training plan is needed",
                    "type": "string",
                    "example": "I need a training plan for improving my freestyle technique"
                },
                "filter": {
                    "description": "Filter allows filtering plans by metadata like difficulty or stroke type",
                    "type": "object",
                    "additionalProperties": {}
                },
                "language": {
                    "description": "Language specifies the language for the response",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Language"
                        }
                    ],
                    "example": "en"
                },
                "method": {
                    "description": "Method can be either 'choose' (select existing plan) or 'generate' (create new plan)",
                    "type": "string",
                    "enum": [
                        "choose",
                        "generate"
                    ],
                    "example": "generate"
                },
                "pool_length": {
                    "description": "PoolLength specifies the pool length for the training plan"
                }
            }
        },
        "models.RAGResponse": {
            "description": "Response containing a generated or selected swim training plan",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "A comprehensive training plan for improving freestyle technique"
                },
                "plan_id": {
                    "description": "PlanID is the identifier of the training plan",
                    "type": "string",
                    "example": "plan_123"
                },
                "table": {
                    "description": "A structured training plan table containing exercise rows",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Row"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Advanced Freestyle Training"
                }
            }
        },
        "models.Row": {
            "description": "A single exercise entry with amount, distance, breaks, content, intensity and total volume",
            "type": "object",
            "properties": {
                "Amount": {
                    "type": "integer",
                    "example": 4
                },
                "Break": {
                    "type": "string",
                    "example": "20"
                },
                "Content": {
                    "type": "string",
                    "example": "Freestyle swim"
                },
                "Distance": {
                    "type": "integer",
                    "example": 100
                },
                "Intensity": {
                    "type": "string",
                    "example": "Easy"
                },
                "Multiplier": {
                    "type": "string",
                    "example": "x"
                },
                "Sum": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "models.UpsertPlanRequest": {
            "description": "Request payload for upserting a swim training plan to the system",
            "type": "object",
            "required": [
                "description",
                "table",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "A comprehensive training plan for improving freestyle technique"
                },
                "plan_id": {
                    "description": "PlanID identifies the training plan to be upserted",
                    "type": "string",
                    "example": "plan_123"
                },
                "table": {
                    "description": "A structured training plan table containing exercise rows",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Row"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Advanced Freestyle Training"
                }
            }
        },
        "models.UpsertPlanResponse": {
            "description": "Response containing the upserted swim training plan",
            "type": "object",
            "properties": {
                "plan_id": {
                    "type": "string",
                    "example": "plan_123"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}